feat: rename agentic-kit to @agentic-kit/chat, add umbrella agentic-kit@2.0.0
