+++ agentic/agent/package.json
@@ -29,7 +29,7 @@
     "test:watch": "jest --watch"
   },
   "dependencies": {
-    "agentic-kit": "workspace:*"
+    "@agentic-kit/chat": "workspace:*"
   },
   "devDependencies": {
     "@agentic-kit/openai": "workspace:*"
