+++ agentic/agentic-kit/README.md
@@ -4,117 +4,30 @@
   <img height="250" src="https://raw.githubusercontent.com/constructive-io/constructive/refs/heads/main/assets/outline-logo.svg" />
 </p>
 
-<p align="center" width="100%">
-  <a href="https://github.com/constructive-io/agentic-kit/actions/workflows/run-tests.yaml">
-    <img height="20" src="https://github.com/constructive-io/agentic-kit/actions/workflows/run-tests.yaml/badge.svg" />
-  </a>
-   <a href="https://github.com/constructive-io/agentic-kit/blob/main/LICENSE"><img height="20" src="https://img.shields.io/badge/license-MIT-blue.svg"/></a>
-   <a href="https://www.npmjs.com/package/agentic-kit"><img height="20" src="https://img.shields.io/github/package-json/v/constructive-io/agentic-kit?filename=packages%2Fagentic-kit%2Fpackage.json"/></a>
-</p>
+Umbrella package for the agentic-kit family — build your own custom agents with a provider-agnostic chat layer, a stateful agent runtime, and the Constructive harness.
 
-A low-level provider portability layer for LLM applications. `agentic-kit`
-provides:
-
-- provider-independent `ModelDescriptor` and `Context` types
-- structured streaming events for text, reasoning, and tool calls
-- model and provider registries
-- cross-provider message normalization for replay and handoff
-- a one-release compatibility wrapper for the legacy `AgentKit.generate()` API
-
-## Installation
-
-```bash
-npm install agentic-kit
-```
-
-
-## Quick Start
-
-### Structured API
+## Usage
 
 ```typescript
-import { complete, getModel } from 'agentic-kit';
+// Chat layer (re-exported at the top level, same API as @agentic-kit/chat)
+import { complete, stream, createOpenAIKit } from 'agentic-kit';
 
-const model = getModel('openai', 'gpt-4o-mini');
-const message = await complete(model!, {
-  messages: [{ role: 'user', content: 'Hello', timestamp: Date.now() }],
-});
+// Agent runtime
+import { agent } from 'agentic-kit';
 
-console.log(message.content);
+// Harness (skills resolution, gating, blueprints)
+import { harness } from 'agentic-kit';
 ```
 
-### Streaming
+## Packages
 
-```typescript
-import { stream, getModel } from 'agentic-kit';
-
-const model = getModel('openai', 'gpt-4o-mini');
-const result = stream(model!, {
-  messages: [{ role: 'user', content: 'Explain tool calling briefly.', timestamp: Date.now() }],
-});
-
-for await (const event of result) {
-  if (event.type === 'text_delta') {
-    process.stdout.write(event.delta);
-  }
-}
-```
-
-## API Reference
-
-### Core API
-
-- `stream(model: ModelDescriptor, context: Context, options?: StreamOptions)`
-- `complete(model: ModelDescriptor, context: Context, options?: StreamOptions)`
-- `completeText(model: ModelDescriptor, context: Context, options?: StreamOptions)`
-
-### Registry
-
-- `registerModel(model)`, `registerModels(models)`, `clearModels()`
-- `registerProvider(provider, sourceId?)`, `unregisterProviders(sourceId)`, `clearProviders()`
-- `getModel(provider, modelId)`, `getModels(provider?)`, `getModelProviders()`
-- `getProvider(api)`, `getRegisteredProviders()`
-
-The package pre-registers `OpenAIAdapter`, `AnthropicAdapter`, and
-`OllamaAdapter` (with empty credentials) plus the built-in model catalogs
-from each adapter package. Re-register an adapter with your own API key to
-activate it.
-
-### Message helpers
-
-- `createUserMessage(content)`, `createAssistantMessage(model)`,
-  `createToolResultMessage(toolCallId, toolName, content, isError?)`
-- `createTextContent(text?)`, `createImageContent(data, mimeType)`,
-  `createToolCall(id, name)`
-- `getMessageText(assistant)` — concatenate text blocks.
-- `cloneMessage(message)` — deep clone.
-- `normalizeContext(context)` — apply per-message normalization.
-- `injectDeferralResults(messages, options?)` — synthesize stand-in
-  `toolResult` messages for every `toolCall` that lacks both a decision and
-  a paired result. Useful when the user types a new message instead of
-  responding to a paused tool — the next request needs a well-formed
-  transcript.
-- `transformMessages(messages, model)` — cross-provider normalization for
-  replay across different providers.
-
-### Adapter re-exports
-
-- `OpenAIAdapter`, `AnthropicAdapter`, `OllamaAdapter`
-- Types: `OpenAIOptions`, `AnthropicOptions`
-- `OllamaClient` for direct Ollama HTTP access.
-
-### Legacy Compatibility API
-
-`AgentKit` is still available for one transition release:
-
-```ts
-interface GenerateInput {
-  model: string;
-  prompt?: string;
-  messages?: Array<{ role: 'system' | 'user' | 'assistant'; content: string }>;
-  stream?: boolean;
-}
-```
-
-Use `createOpenAIKit`, `createAnthropicKit`, `createOllamaKit`, or
-`createMultiProviderKit()` if you still need the old prompt-only entrypoint.
+| Package | Description |
+|---------|-------------|
+| [`@agentic-kit/chat`](https://github.com/constructive-io/constructive/tree/main/agentic/chat) | Provider-agnostic LLM adapter with streaming and multi-turn support |
+| [`@agentic-kit/protocol`](https://github.com/constructive-io/constructive/tree/main/agentic/protocol) | Shared types and event stream |
+| [`@agentic-kit/openai`](https://github.com/constructive-io/constructive/tree/main/agentic/openai) | OpenAI-compatible provider adapter |
+| [`@agentic-kit/anthropic`](https://github.com/constructive-io/constructive/tree/main/agentic/anthropic) | Anthropic provider adapter |
+| [`@agentic-kit/ollama`](https://github.com/constructive-io/constructive/tree/main/agentic/ollama) | Ollama provider adapter |
+| [`@agentic-kit/agent`](https://github.com/constructive-io/constructive/tree/main/agentic/agent) | Minimal stateful agent runtime |
+| [`@agentic-kit/harness`](https://github.com/constructive-io/constructive/tree/main/agentic/harness) | Host-neutral coding-agent harness: skills layers, confirm gating, blueprints |
+| [`@agentic-kit/react`](https://github.com/constructive-io/constructive/tree/main/agentic/react) | React bindings (`useChat`) — install separately (React peer dependency) |
