+++ agentic/chat/__tests__/adapter.test.ts
