+++ agentic/chat/__tests__/helpers.ts
