+++ agentic/chat/__tests__/inject-deferral-results.test.ts
