+++ agentic/chat/src/model-registry.ts
