+++ agentic/chat/src/provider-registry.ts
