+++ agentic/chat/src/transform-messages.ts
